# api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Examify Swagger",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/api/bonus/get-bonus-information-me": {
            "get": {
                "description": "Retrieve bonus information by the current user's chat ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bonus"
                ],
                "summary": "Get Bonus Information for Current User",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.GetBonusInformationByChatIdResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pb.GetBonusInformationByChatIdResponse": {
            "type": "object",
            "properties": {
                "bonusCount": {
                    "type": "integer"
                },
                "more": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.GetMoreBonusInformation"
                    }
                },
                "name": {
                    "type": "string"
                },
                "refLink": {
                    "type": "string"
                },
                "registeredAt": {
                    "type": "string"
                },
                "welcomeCount": {
                    "type": "integer"
                }
            }
        },
...
        "pb.GetMoreBonusInformation": {
            "type": "object",
            "properties": {
                "guestChatId": {
                    "type": "string"
                },
                "guestName": {
                    "type": "string"
                },
                "guestRegisteredAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}